Update Registro de Propietarios

+++ FrontEnd/CanchaLibre/CanchaLibreWeb/CanchaLibreWeb/ViewModels/PropietarioViewModel.cs
@@ -4,7 +4,7 @@ namespace CanchaLibreWeb.ViewModels;
 
 public class PropietarioViewModel : PersonaViewModel
 {
-    public int Calificacion {get; set;}
+    public double Calificacion {get; set;}
 
     [Required(ErrorMessage = "Debe ingresar su RUC")]
     public string Ruc { get; set; } = string.Empty;
